Move init command files out of a separate package
